fix: stabilize dashboard hmr and align stock helper tests
